ui: add a basic progress bar to the player
In our current design, MediaMod polls services every 3 seconds to check if
there is any new track information. Because of this, the progress bar needs
to "guess" what the current progress is in the time between each poll. It
does this by storing the time when the track is set, and then on each frame,
it will take the difference between the current time and the old time, and
then add that to the elapsed amount that we last received.

TL;DR: The progress bar is smooth, even if we are polling every 3 seconds
for a new track :)