feat: add a basic theming system
This allows the user to customise the appearance of the MediaMod Player HUD.
Right now, themes can be created by making a new file in `.minecraft/mediamod/themes`.
At the moment, you can set the colour for the background, text, progress bar and the
progress bar's background.

An example theme can be seen below:
{
    "name": "Custom",
    "colors": {
        "background": "#00",
        "text": "#A4A4A6",
        "progress_bar": "#96D294",
        "progress_bar_background": "#58595B"
    }
}