feat: implement a basic oauth "callback" server
Once the user clicks the "Login" button, and approves it when their
browser opens, Spotify will redirect them to our callback server.
The callback server resides at localhost:9103. Its purpose is to
send the code that we receive from Spotify to our API for a code to
access token exchange.