feat: add a basic 'dynamic' theme
This theme will take its colours from the album art, it still needs some
work, but it's a start!

The only problem at the moment is, the theme's colours only change when
the track is changed, we don't have a way to get the current album art
when the theme is being switched.