feat(ui): show the theme name when selected in the theme editor
This is the start of the implementation of the actual theme editing
functionality, we have a ThemeEditorContainer which has a state for
the current theme, when that's set, we'll update our UI to reflect
that theme, allowing the user to edit the theme.

Right now, the UI only shows the selected theme's name. The left column
also now highlights the theme that's currently selected.