feat: warmup kreflect on startup
On certain machines, kreflect can have a pretty large performance impact
on startup. To mitigate this, we do a dummy kreflect call on another
thread when starting the game in order to not freeze the game for a
second or two when its being initialized later on.

NOTE: We don't do this on 1.8.9 or 1.12.2 as Essential does it for us.